Move c3.js to vendor/assets/ and serve separately
The vendor/assets/ directory is a more appropriate place for this file:

http://guides.rubyonrails.org/asset_pipeline.html#asset-organization

> app/assets is for assets that are owned by the application, such as
> custom images, JavaScript files or stylesheets.

> vendor/assets is for assets that are owned by outside entities, such
> as code for JavaScript plugins and CSS frameworks.

Code Climate is currently analysing this file as application code, and
the results are dragging the project GPA down to a hilariously low 0.2:

https://codeclimate.com/github/intercom/pulse/app/assets/javascripts/c3.js

One consequence of moving this file out of app/assets/ is that it isn't
automatically included in application.js any more. Adding it back in is
as easy as `//= require d3`, but I think it's actually better to serve
it separately, since then users won't have to download the same version
of the C3 library again every time the application's JavaScript changes.